lal.series:  Param instance's .pcdata is already in native type - no need for type cast Original: f6199f9db50adb0af61d912faf68befc1ea3299f

+++ lal/python/lal/series.py
@@ -97,7 +97,7 @@ def _parse_series(elem, creatorfunc, delta_target_unit_string):
     series = creatorfunc(
         str(a.Name),
         epoch,
-        float(f0.pcdata) * float(f0_unit / inverse_seconds_unit),
+        f0.pcdata * float(f0_unit / inverse_seconds_unit),
         dims[0].Scale * float(delta_unit / delta_target_unit),
         sample_unit,
         len(a.array.T)
